+++ CHANGELOG.md
@@ -60,6 +60,7 @@
 - Stock Move: Do not modify wap when generating customer return.
 - ADDRESS: Fix error message when clicking on ViewMap Btn of a new address.
 - Configurator BOM: Correctly make the difference between components and sub bill of material.
+- Advance import: Fix multiple search on same field
 
 ## [5.2.3] - 2020-01-23
 ## Features